Create Watchdog-Timer-RISC-V.v
This module has five inputs: clk, rst, en, timeout, and an output wdt. The clk input is the clock signal that drives the timer, rst is the reset signal, en is the enable signal, and timeout is the value that the counter cnt will be compared against, to decide if the watchdog timer has expired. The wdt output is a signal that indicates when the timer has expired.

The module uses an internal 32-bit counter cnt that is incremented on each rising edge of the clk signal, as long as the en signal is high. The counter is reset to 0 when the rst signal is high.

When the counter reaches its maximum value timeout, the wdt signal is set to 1 and the counter is reset to 0. Otherwise, the wdt signal is set to 0.

And here's an example of a testbench for the above watchdog timer: