Create Testbench-Watchdog-Timer-RISC-V.v
In the testbench, I created all the necessary input regs, clk, rst, en and timeout, and wire for the output, wdt. Then instantiated the module under test (uut) with these inputs and output.
In the initial block, I set the initial values for inputs, rst and en are high, and clk starts with 0, and timeout is set to 32'h3fffffff. Then I set rst to low after 5 ns and en to 1 after another 5 ns. Finally, I created an infinite loop to toggle the clock signal every 5 ns.